rearranged the files in cli folder for better comprehension
