+++ ether_sql/cli/__init__.py
@@ -1,4 +1,5 @@
 import click
+from ether_sql.cli import database
 
 
 @click.group()
@@ -6,7 +7,7 @@ def cli():
     """CLI script for ether_sql"""
 
 
-# cli.add_command(cli.manager, "database")
+cli.add_command(database.cli, "database")
 
 
 @cli.command()
@@ -15,23 +16,3 @@ def check_settings():
     import settings
     for name, item in settings.all_settings().iteritems():
         print("{} = {}".format(name, item))
-
-
-@cli.command()
-def create_tables():
-    """Create the database tables."""
-    from ether_sql.models import db
-    from ether_sql import session
-
-    db.metadata.create_all(session)
-    click.echo('Created the tables')
-
-
-@cli.command()
-def drop_tables():
-    """Drop the database tables."""
-    from ether_sql.models import db
-    from ether_sql import session
-
-    db.metadata.drop_all(session)
-    click.echo('Dropped the tables')
